Fixes the checks of existing layouts
With new json backend, the way to autodetect layouts
has changed. It tries looking for .json files and then
for .xml ones, but if the configuration has the suffix
already, then fails and do not load anything.

Now, it first checks if the configured layout exists.
If not, then it fallbacks to the standard autodetection

https://bugzilla.gnome.org/show_bug.cgi?id=622221