bin: fix path settings to never include empty path elements. fix typos
Things like PYTHONPATH="foo:$PYTHONPATH" are bad, because if
PYTHONPATH isn't previously set, you'll end up with a "" element in
it, which will be interpreted as ".", which you don't want.

Use some tricky bash syntax to say "add : and $PYTHONPATH if
PYTHONPATH is set".

Also fix a few typos ("licaribou", etc)

https://bugzilla.gnome.org/show_bug.cgi?id=651005