daemon: Start with D-Bus activation
Install the daemon to libexecdir. Make the daemon own a name on the
session bus, and install a D-Bus service file for it. Allow the name
to be replaced by another process, so that a D-Bus activated daemon can
be replaced by a GSettings-key autostarted daemon. Handle run() being
called multiple times.

https://bugzilla.gnome.org/show_bug.cgi?id=683712