Add label to all keys
Some keys do not have a label, so one has to guess what it does.

Do the following for keys we do not have a label yet.
First, try to find a label for dead keys by looking for the label
of the non-dead key.
Second, if no label is found use the key name as a label.

https://bugzilla.gnome.org/show_bug.cgi?id=739526