Allow to overwrite the label in the keyboard layout.
Currently some buttons do only have its name as label.
This is not always useful. I.e. for the dead-keys the label is often only "dead".
In some keyboard layout it is preferred to have a symbol as label.

This allows to overwrite the default label in the keyboard layout.

https://bugzilla.gnome.org/show_bug.cgi?id=739837